doing initial load, will need to add data folder as needed.
Install Node.JS for your system.
Navigate to the containing folder for this project
do a npm install
install mongodb
in a separate terminal window, start mongodb to the test database with…
./mongod --dbpath ~/Documents/Code/photon-server/data

Then, you can ./mongo into the database as needed,
then if you run
node app.js
then it will be running and you can test adding and subtracting users
and all that good stuff.